1. Moving AST printing functions from parser module to AST module. 2. Starting skeleton for type checker. 3. Continuing to work on hash table. 4. Adding test case. 5. Fixing function call parsing bug (found with above test case).
