+++ src/error.h
@@ -9,7 +9,7 @@
 
 // 2 pages worth of debug message buffer
 #define MAX_DEBUG_LEN 4096 * 2
-#define MAX_ERR_LEN 1024
+#define MAX_ERROR_LEN 1024
 
 enum exit_error_types {
     LEXER_ERROR_UNKNOWN_CHARACTER = 100,
