+++ test/tcfunction.lb
@@ -1,0 +1,18 @@
+' Test Name: Type Check Function
+' Test Category: Type Checking
+
+func main(string[] args) -> int
+then
+    string greeting := "My name is Liam. What is your name?";
+    string answer;
+    answer := getstdin();
+
+    if (getstdin() == "Jim") then
+        println("Hello Jim");
+    else then
+        println("Hello " + answer);
+    end
+
+    return 0;
+end
+
